+++ app/build.gradle
@@ -110,8 +110,8 @@ jacocoTestCoverageVerification {
 
             excludes = [
                     "com.codesoom.assignment.App",
-                    "com.codesoom.assignment.dto.*",
-                    "com.codesoom.assignment.domain.*"
+                    "com.codesoom.assignment.product.dto.*",
+                    "com.codesoom.assignment.product.domain.*"
             ]
         }
     }
