+++ app/src/main/java/com/codesoom/assignment/HelloController.java
@@ -1,0 +1,12 @@
+package com.codesoom.assignment;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+    @RequestMapping("/")
+    public String sayHello() {
+        return "Hello, world!";
+    }
+}
