--- app/src/main/java/com/codesoom/assignment/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
-        super("Product not found: " + id);
-    }
-}
