+++ app/src/main/java/com/codesoom/assignment/common/exceptions/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.common.exceptions;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(Long id) {
+        super("Product not found: " + id);
+    }
+}
