+++ app/src/main/java/com/codesoom/assignment/common/exceptions/UserNotFoundException.java
@@ -1,0 +1,13 @@
+package com.codesoom.assignment.common.exceptions;
+
+public class UserNotFoundException extends RuntimeException {
+
+    public UserNotFoundException() {
+        super("회원을 찾을 수 없습니다.");
+    }
+
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+
+}
