--- app/src/main/java/com/codesoom/assignment/controllers/NotFoundErrorAdvice.java
@@ -1,29 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-import com.codesoom.assignment.ProductNotFoundException;
-import com.codesoom.assignment.UserNotFoundException;
-import com.codesoom.assignment.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ControllerAdvice
-public class NotFoundErrorAdvice {
-
-    @ResponseBody
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ErrorResponse handleProductTaskNotFound() {
-        return new ErrorResponse("Product not found");
-    }
-
-    @ResponseBody
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(UserNotFoundException.class)
-    public ErrorResponse handleUserNotFound(UserNotFoundException exception) {
-        return new ErrorResponse(exception.getMessage());
-    }
-
-}
