--- app/src/main/java/com/codesoom/assignment/controllers/UserController.java
@@ -1,60 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-import com.codesoom.assignment.application.UserService;
-import com.codesoom.assignment.dto.UserCreateRequest;
-import com.codesoom.assignment.dto.UserResponse;
-import com.codesoom.assignment.dto.UserUpdateRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.validation.Valid;
-import java.util.List;
-
-@RestController
-@CrossOrigin
-@RequestMapping("/users")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    @GetMapping
-    public List<UserResponse> list() {
-        return userService.getUsers();
-    }
-
-    @GetMapping("{id}")
-    public UserResponse find(@PathVariable Long id) {
-        return userService.getUser(id);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserResponse create(@RequestBody @Valid UserCreateRequest createRequest) {
-        return userService.createUser(createRequest);
-    }
-
-    @PatchMapping("{id}")
-    public UserResponse update(
-            @PathVariable Long id,
-            @RequestBody @Valid UserUpdateRequest updateRequest) {
-        return userService.updateUser(id, updateRequest);
-    }
-
-    @DeleteMapping("{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Long id) {
-        userService.deleteUser(id);
-    }
-
-}
