--- app/src/main/java/com/codesoom/assignment/domain/User.java
@@ -1,42 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String name;
-
-    private String email;
-
-    private String password;
-
-    @Builder
-    private User(Long id, String name, String email, String password) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-        this.password = password;
-    }
-
-    public User changeWith(User source) {
-        this.name = source.name;
-        this.password = source.password;
-
-        return this;
-    }
-
-}
