--- app/src/main/java/com/codesoom/assignment/domain/UserRepository.java
@@ -1,16 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface UserRepository {
-
-    List<User> findAll();
-
-    Optional<User> findById(Long id);
-
-    User save(User user);
-
-    void delete(User user);
-
-}
