--- app/src/main/java/com/codesoom/assignment/dto/UserUpdateRequest.java
@@ -1,29 +1,0 @@
-package com.codesoom.assignment.dto;
-
-import com.github.dozermapper.core.Mapping;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserUpdateRequest {
-
-    @NotBlank
-    @Mapping("name")
-    private String name;
-
-    @NotBlank
-    @Mapping("password")
-    private String password;
-
-    @Builder
-    public UserUpdateRequest(String name, String password) {
-        this.name = name;
-        this.password = password;
-    }
-
-}
