+++ app/src/main/java/com/codesoom/assignment/product/controller/ProductController.java
@@ -1,0 +1,60 @@
+package com.codesoom.assignment.product.controller;
+
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.dto.ProductData;
+import com.codesoom.assignment.product.service.ProductService;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+import java.util.List;
+
+@RestController
+@CrossOrigin
+@RequestMapping("/products")
+public class ProductController {
+    private final ProductService productService;
+
+    public ProductController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping
+    public List<Product> list() {
+        return productService.getProducts();
+    }
+
+    @GetMapping("{id}")
+    public Product detail(@PathVariable Long id) {
+        return productService.getProduct(id);
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public Product create(@RequestBody @Valid ProductData productData) {
+        return productService.createProduct(productData);
+    }
+
+    @PatchMapping("{id}")
+    public Product update(
+            @PathVariable Long id,
+            @RequestBody @Valid ProductData productData
+    ) {
+        return productService.updateProduct(id, productData);
+    }
+
+    @DeleteMapping("{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void destroy(@PathVariable Long id) {
+        productService.deleteProduct(id);
+    }
+}
