+++ app/src/main/java/com/codesoom/assignment/product/domain/Product.java
@@ -1,0 +1,38 @@
+package com.codesoom.assignment.product.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Product {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String name;
+
+    private String maker;
+
+    private Integer price;
+
+    private String imageUrl;
+
+    public Product changeWith(Product source) {
+        this.name = source.name;
+        this.maker = source.maker;
+        this.price = source.price;
+        this.imageUrl = source.imageUrl;
+
+        return this;
+    }
+}
