+++ app/src/main/java/com/codesoom/assignment/user/infra/JpaUserRepository.java
@@ -1,0 +1,21 @@
+package com.codesoom.assignment.user.infra;
+
+import com.codesoom.assignment.user.domain.User;
+import com.codesoom.assignment.user.domain.UserRepository;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface JpaUserRepository
+        extends UserRepository, CrudRepository<User, Long> {
+
+    List<User> findAll();
+
+    Optional<User> findById(Long id);
+
+    User save(User user);
+
+    void delete(User user);
+
+}
