--- app/src/test/java/com/codesoom/assignment/controllers/HelloControllerTest.java
@@ -1,14 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class HelloControllerTest {
-    @Test
-    void sayHello() {
-        HelloController controller = new HelloController();
-
-        assertThat(controller.sayHello()).isEqualTo("Hello, world!");
-    }
-}
